Ensure correct `eslint` version is being used.
As of 4.0.0 we require `eslint@^8.4.0` (and properly included that as a
non-optional peer dependency), but unfortunately (depending on your
package manager) the peer dependency isn't enforced.

This brings in
[validate-peer-dependencies](https://github.com/rwjblue/validate-peer-dependencies)
and runs it when the formatter is required/loaded. This will emit a
helpful error if your `eslint` version isn't compliant with our peer
dependency requirements.